requirements_dev.txt
-r requirements.txt

ipython

django-debug-toolbar
django-extensions
